refactor(catalog): split export pipeline module
